Containerize application with Docker
Docker can be used to containerize the frontend and backend in order
to standardize setting up their environments and running each service.

An NGINX reverse proxy was set up so that both services are served from
the domain (localhost). This avoids CORS-related issues that we were
previously solving using JWT tokens. These tokens are no longer required
in favour of Django's session authentication which is included in this
commit.