Create celery beat task for clearing old guests
Celery beat is a scheduler for running tasks at regular intervals. It's
used here to run the clear old guests command periodically (currently
set to run each minute for testing purposes).

I've also created a docker image for the postgres database. This is
obviously not suited for production but it makes development a bit
easier.

The way the image is configured in `docker-compose.yaml` creates a
`data/` directory so I've added said directory to the gitignore.